fastapi-solution/requirements/base.txt
redis==4.4.2
elasticsearch[async]==7.9.1
fastapi==0.95.2
fastapi-cache2[redis]==v0.2.1
gunicorn==20.1.0
httptools==0.5.0
orjson==3.8.13
pydantic==1.9.0
uvicorn==0.12.2
uvloop==0.17.0 ; sys_platform != "win32" and implementation_name == "cpython"
